Update README.md
Face recognition is the problem of identifying and verifying people in a photograph by their face. It is
a task that is trivially performed by humans, even under varying light and when faces are changed by
age or obstructed with accessories and facial hair. Nevertheless, it has remained a challenging
computer vision problem for decades until recently.